+++ opencascade-hs/src/OpenCascade/HLRBRep/Internal/Destructors.hs
@@ -2,6 +2,7 @@
 module OpenCascade.HLRBRep.Internal.Destructors 
 ( deleteAlgo
 , deleteHandleAlgo
+, deleteHLRToShape
 ) where
 
 import OpenCascade.HLRBRep.Types
@@ -11,4 +12,4 @@ import Foreign.Ptr
 
 foreign import capi unsafe "hs_HLRBRep_Algo.h hs_delete_HLRBRep_Algo" deleteAlgo :: Ptr Algo -> IO ()
 foreign import capi unsafe "hs_HLRBRep_Algo.h hs_delete_Handle_HLRBRep_Algo" deleteHandleAlgo :: Ptr (Handle Algo) -> IO ()
-
+foreign import capi unsafe "hs_HLRBRep_HLRToShape.h hs_delete_HLRBRep_HLRToShape" deleteHLRToShape :: Ptr HLRToShape -> IO ()