+++ opencascade-hs/src/OpenCascade/HLRBRep/TypeOfResultingEdge.hs
@@ -1,0 +1,12 @@
+module OpenCascade.HLRBRep.TypeOfResultingEdge
+( TypeOfResultingEdge (..)
+) where 
+
+data TypeOfResultingEdge 
+    = Undefined
+    | IsoLine
+    | OutLine
+    | Rg1Line
+    | RgNLine
+    | Sharp
+    deriving (Eq, Enum, Show)