+++ opencascade-hs/src/OpenCascade/HLRBRep/Types.hs
@@ -1,6 +1,8 @@
 {-# LANGUAGE EmptyDataDecls #-}
 module OpenCascade.HLRBRep.Types
 ( Algo
+, HLRToShape
 ) where
 
 data Algo
+data HLRToShape